Moved slot methods to mainwindow.cpp
