+++ qtEditor/mainwindow.cpp
@@ -8,6 +8,7 @@
 #include <iostream>
 #include "ImportImageDialog.h"
 #include "ExportImageDialog.h"
+#include "aboutdialog.h"
 #include <QMessageBox>
 #include <QFileDialog>
 #include <QInputDialog>
@@ -481,6 +482,76 @@ void MainWindow::openExportImageDialog()
     delete exportImageDialog;
 }
 
+void MainWindow::importImage()
+{
+    openImportImageDialog();
+}
+
+void MainWindow::exportImage()
+{
+    openExportImageDialog();
+}
+
+void MainWindow::pencilEraserGPU()
+{
+    selectTool("HeightfieldOperationGPUPencil", ScapeEngine::SCAPEUIELEMENTGROUPID_OPERATION);
+}
+
+void MainWindow::directionalNoiseGPU()
+{
+    selectTool("HeightfieldOperationGPUNoiseDrag", ScapeEngine::SCAPEUIELEMENTGROUPID_OPERATION);
+}
+
+void MainWindow::noiseGPU()
+{
+    selectTool("HeightfieldOperationGPUNoise", ScapeEngine::SCAPEUIELEMENTGROUPID_OPERATION);
+}
+
+void MainWindow::softenGPU()
+{
+    selectTool("HeightfieldOperationGPUBrush", ScapeEngine::SCAPEUIELEMENTGROUPID_OPERATION);
+}
+
+void MainWindow::sharpenSoften()
+{
+    selectTool("HeightfieldOperationCPUSmooth", ScapeEngine::SCAPEUIELEMENTGROUPID_OPERATION);
+}
+
+void MainWindow::noiseCPU()
+{
+    selectTool("HeightfieldOperationCPUNoise", ScapeEngine::SCAPEUIELEMENTGROUPID_OPERATION);
+}
+
+void MainWindow::levelUnlevel()
+{
+    selectTool("HeightfieldOperationCPULevel", ScapeEngine::SCAPEUIELEMENTGROUPID_OPERATION);
+}
+
+void MainWindow::waterErosion()
+{
+    selectTool("HeightfieldOperationCPUErosion", ScapeEngine::SCAPEUIELEMENTGROUPID_OPERATION);
+}
+
+void MainWindow::pencilEraser()
+{
+    selectTool("HeightfieldOperationCPUBrush", ScapeEngine::SCAPEUIELEMENTGROUPID_OPERATION);
+}
+
+void MainWindow::materialProps()
+{
+    selectTool("MaterialSettings", ScapeEngine::SCAPEUIELEMENTGROUPID_MATERIAL);
+}
+
+void MainWindow::skyProps()
+{
+    selectTool("SkySettings", ScapeEngine::SCAPEUIELEMENTGROUPID_SKY);
+}
+
+void MainWindow::renderwindowProps()
+{
+    selectTool("RenderWindowSettings", ScapeEngine::SCAPEUIELEMENTGROUPID_RENDERWINDOW);
+}
+
 void MainWindow::createPreset()
 {
     bool ok;
@@ -558,6 +629,15 @@ void MainWindow::importPreset()
         populatePresetPanel();
     }
 }
+
+void MainWindow::aboutApp()
+{
+    AboutDialog dialog(this);
+    if (dialog.exec() == QDialog::Accepted)
+        return;
+}
+
+void MainWindow::exitApp() { close(); }
 
 void MainWindow::propertyValueChanged(const std::string& key, const std::string& value)
 {
